Moved complex dot tests to intermediate class
